Added python 3.0.x, 3.1.x, 3.2.x, 3.3.x, and 3.4.x.
As reported in https://github.com/actions/setup-python/issues/157 these python versions was missing on windows to be able to basically use python code requiring these odd versions of python, some of which builds extension modules that is then needed for pypi packages for example.

This is needed because what if they want to support those python versions even though they are end of life anyway? Linux has some of these versions so sometimes linux does not apply but, I will need help for those with linux to actually build the things needed to add linux and mac to these versions in this pull request.

Also will file an issue in virtual-environments repository as well as this is insane.