fix: release assets upload
actions/python-versions#267 introduced a regression in asset upload.

The layout of the artefacts to upload has changed. They're now all flatten in the working directory whereas before that PR, only the `hashes.sha256` file was directly in the working directory.
The change in layout caused `.toString()` to be called on python archives which corrupts the data that gets uploaded (as can be seen by comparing sha256 in `hashes.sha256` with the ones from other uploaded assets).

This commit ensures that a "flat layout" is used and that `toString()` is only called for `hashes.sha256`.